doghouse: separate checker to new file
