+++ doghouse/appengine/main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"log"
@@ -10,12 +9,9 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/haya14busa/reviewdog/doghouse"
-	"github.com/haya14busa/reviewdog/doghouse/server"
 	"github.com/haya14busa/reviewdog/doghouse/server/cookieman"
 	"github.com/haya14busa/secretbox"
 	"google.golang.org/appengine"
-	"google.golang.org/appengine/urlfetch"
 )
 
 func mustCookieMan() *cookieman.CookieMan {
@@ -101,51 +97,4 @@ func main() {
 
 func handleTop(w http.ResponseWriter, r *http.Request) {
 	fmt.Fprintln(w, "reviewdog")
-}
-
-type githubChecker struct {
-	privateKey    []byte
-	integrationID int
-}
-
-func (gc *githubChecker) handleCheck(w http.ResponseWriter, r *http.Request) {
-	if r.Method != "POST" {
-		w.WriteHeader(http.StatusMethodNotAllowed)
-		return
-	}
-	var req doghouse.CheckRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		fmt.Fprintf(w, "failed to decode request: %v", err)
-		return
-	}
-	ctx := appengine.NewContext(r)
-
-	opt := &server.NewGitHubClientOption{
-		PrivateKey:     gc.privateKey,
-		IntegrationID:  gc.integrationID,
-		InstallationID: req.InstallationID,
-		RepoOwner:      req.Owner,
-		RepoName:       req.Repo,
-		Client:         urlfetch.Client(ctx),
-	}
-
-	gh, err := server.NewGitHubClient(ctx, opt)
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		fmt.Fprintln(w, err)
-		return
-	}
-
-	res, err := server.NewChecker(&req, gh).Check(ctx)
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		fmt.Fprintln(w, err)
-		return
-	}
-	if err := json.NewEncoder(w).Encode(res); err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		fmt.Fprintln(w, err)
-		return
-	}
 }
